Refactor: made some changes on layout and default page tsx's, and renamed nav-bar folder
