+++ src/app/layout.tsx
@@ -45,7 +45,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={`${metrapolis.variable} font-metrapolis`}>
+      <body className={`${metrapolis.variable}`}>
         {children}
         <Analytics />
       </body>
