+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import HeroBanner from "@/components/ui/main-section/hero-banner";
 import WeekendOffers from "@/components/ui/main-section/weekend-offers";
-import NavbarContent from "@/components/ui/nav-bar/nav-bar-content";
+import NavbarContent from "@/components/ui/nav-bar-section/nav-bar-content";
 
 export default function Home() {
   return (
